validator_worker - mod worker - move fns for running worker
